Record and rollback keys in a test
This change intends to make it easier to override the default wiring by
the redis service.  Primarily this will be useful for testing.

The RedisService is now in src, but is wired manually in
RedisGrailsPlugin.  Some of the wiring methods that were in
RedisGrailsPlugin have been moved to RedisConfigurationUtil so they can
be more easily accessed by apps using the plugin.